fix : s3 url 미존재시 에러처리

+++ src/main/java/com/app/pingpong/domain/image/S3Uploader.java
@@ -5,6 +5,7 @@ import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
+import com.app.pingpong.global.common.exception.BaseException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
@@ -17,6 +18,8 @@ import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
+
+import static com.app.pingpong.global.common.exception.StatusCode.INVALID_S3_URL;
 
 
 @Service
@@ -57,7 +60,12 @@ public class S3Uploader {
     }
 
     public void deleteFile(String fileName) {
-        amazonS3.deleteObject(new DeleteObjectRequest(bucket, fileName));
+        if (amazonS3.doesObjectExist(bucket, fileName)) {
+            amazonS3.deleteObject(new DeleteObjectRequest(bucket, fileName));
+        } else {
+
+            throw new BaseException(INVALID_S3_URL);
+        }
     }
 
     public String getFilePath(String path) {
